+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v2.1.0](https://github.com/allenai/beaker-gantry/releases/tag/v2.1.0) - 2025-05-08
+
 ### Added
 
 - Added `--all` flag to `gantry list` to search all experiments, not just ones submitted through Gantry.